Fixed type error in VertexUtilityTest

+++ src/test/VertexUtilityTest.cpp
@@ -28,10 +28,10 @@ BOOST_AUTO_TEST_SUITE(VertexUtilityTest)
  */
 BOOST_FIXTURE_TEST_CASE(GetPosition_IndexOutOfBounds, VertexUtility)
 {
-    Point<int> v = VertexUtility::GetPosition(10, 2, 2);
+    Point<uint16_t> v = VertexUtility::GetPosition(10, 2, 2);
         
-    BOOST_REQUIRE_EQUAL(v.x, 0);
-    BOOST_REQUIRE_EQUAL(v.y, 5);
+    BOOST_REQUIRE_EQUAL(v.x, 0u);
+    BOOST_REQUIRE_EQUAL(v.y, 5u);
 }
 
 /**
@@ -40,10 +40,10 @@ BOOST_FIXTURE_TEST_CASE(GetPosition_IndexOutOfBounds, VertexUtility)
  */
 BOOST_FIXTURE_TEST_CASE(GetPosition_CornerIndex, VertexUtility)
 {
-    Point<int> v = VertexUtility::GetPosition(8, 3, 3);
+    Point<uint16_t> v = VertexUtility::GetPosition(8, 3, 3);
         
-    BOOST_REQUIRE_EQUAL(v.x, 2);
-    BOOST_REQUIRE_EQUAL(v.y, 2);
+    BOOST_REQUIRE_EQUAL(v.x, 2u);
+    BOOST_REQUIRE_EQUAL(v.y, 2u);
 }
     
 /**
@@ -56,9 +56,9 @@ BOOST_FIXTURE_TEST_CASE(GetPosition_ZeroIndex, VertexUtility)
     {
         for (int height = 1; height < 100; height++)
         {
-            Point<int> v = VertexUtility::GetPosition(0, width, height);
-            BOOST_REQUIRE_EQUAL(v.x, 0);
-            BOOST_REQUIRE_EQUAL(v.y, 0);
+            Point<uint16_t> v = VertexUtility::GetPosition(0, width, height);
+            BOOST_REQUIRE_EQUAL(v.x, 0u);
+            BOOST_REQUIRE_EQUAL(v.y, 0u);
         }
     }
 }
